Enum transaction_builder_generated::stdlib::ScriptCall

pub enum ScriptCall {
    AddCurrencyToAccount {
        currency: TypeTag,
    },
    AddRecoveryRotationCapability {
        recovery_address: AccountAddress,
    },
    AddToScriptAllowList {
        hash: Vec<u8>,
        sliding_nonce: u64,
    },
    AddValidatorAndReconfigure {
        sliding_nonce: u64,
        validator_name: Vec<u8>,
        validator_address: AccountAddress,
    },
    Burn {
        token: TypeTag,
        sliding_nonce: u64,
        preburn_address: AccountAddress,
    },
    BurnTxnFees {
        coin_type: TypeTag,
    },
    CancelBurn {
        token: TypeTag,
        preburn_address: AccountAddress,
    },
    CreateChildVaspAccount {
        coin_type: TypeTag,
        child_address: AccountAddress,
        auth_key_prefix: Vec<u8>,
        add_all_currencies: bool,
        child_initial_balance: u64,
    },
    CreateDesignatedDealer {
        currency: TypeTag,
        sliding_nonce: u64,
        addr: AccountAddress,
        auth_key_prefix: Vec<u8>,
        human_name: Vec<u8>,
        add_all_currencies: bool,
    },
    CreateParentVaspAccount {
        coin_type: TypeTag,
        sliding_nonce: u64,
        new_account_address: AccountAddress,
        auth_key_prefix: Vec<u8>,
        human_name: Vec<u8>,
        add_all_currencies: bool,
    },
    CreateRecoveryAddress {},
    CreateValidatorAccount {
        sliding_nonce: u64,
        new_account_address: AccountAddress,
        auth_key_prefix: Vec<u8>,
        human_name: Vec<u8>,
    },
    CreateValidatorOperatorAccount {
        sliding_nonce: u64,
        new_account_address: AccountAddress,
        auth_key_prefix: Vec<u8>,
        human_name: Vec<u8>,
    },
    FreezeAccount {
        sliding_nonce: u64,
        to_freeze_account: AccountAddress,
    },
    MintLbr {
        amount_lbr: u64,
    },
    PeerToPeerWithMetadata {
        currency: TypeTag,
        payee: AccountAddress,
        amount: u64,
        metadata: Vec<u8>,
        metadata_signature: Vec<u8>,
    },
    Preburn {
        token: TypeTag,
        amount: u64,
    },
    PublishSharedEd25519PublicKey {
        public_key: Vec<u8>,
    },
    RegisterValidatorConfig {
        validator_account: AccountAddress,
        consensus_pubkey: Vec<u8>,
        validator_network_identity_pubkey: Vec<u8>,
        validator_network_address: Vec<u8>,
        fullnodes_network_identity_pubkey: Vec<u8>,
        fullnodes_network_address: Vec<u8>,
    },
    RemoveValidatorAndReconfigure {
        sliding_nonce: u64,
        validator_name: Vec<u8>,
        validator_address: AccountAddress,
    },
    RotateAuthenticationKey {
        new_key: Vec<u8>,
    },
    RotateAuthenticationKeyWithNonce {
        sliding_nonce: u64,
        new_key: Vec<u8>,
    },
    RotateAuthenticationKeyWithNonceAdmin {
        sliding_nonce: u64,
        new_key: Vec<u8>,
    },
    RotateAuthenticationKeyWithRecoveryAddress {
        recovery_address: AccountAddress,
        to_recover: AccountAddress,
        new_key: Vec<u8>,
    },
    RotateDualAttestationInfo {
        new_url: Vec<u8>,
        new_key: Vec<u8>,
    },
    RotateSharedEd25519PublicKey {
        public_key: Vec<u8>,
    },
    SetValidatorConfigAndReconfigure {
        validator_account: AccountAddress,
        consensus_pubkey: Vec<u8>,
        validator_network_identity_pubkey: Vec<u8>,
        validator_network_address: Vec<u8>,
        fullnodes_network_identity_pubkey: Vec<u8>,
        fullnodes_network_address: Vec<u8>,
    },
    SetValidatorOperator {
        operator_name: Vec<u8>,
        operator_account: AccountAddress,
    },
    SetValidatorOperatorWithNonceAdmin {
        sliding_nonce: u64,
        operator_name: Vec<u8>,
        operator_account: AccountAddress,
    },
    TieredMint {
        coin_type: TypeTag,
        sliding_nonce: u64,
        designated_dealer_address: AccountAddress,
        mint_amount: u64,
        tier_index: u64,
    },
    UnfreezeAccount {
        sliding_nonce: u64,
        to_unfreeze_account: AccountAddress,
    },
    UnmintLbr {
        amount_lbr: u64,
    },
    UpdateDualAttestationLimit {
        sliding_nonce: u64,
        new_micro_lbr_limit: u64,
    },
    UpdateExchangeRate {
        currency: TypeTag,
        sliding_nonce: u64,
        new_exchange_rate_numerator: u64,
        new_exchange_rate_denominator: u64,
    },
    UpdateLibraVersion {
        sliding_nonce: u64,
        major: u64,
    },
    UpdateMintingAbility {
        currency: TypeTag,
        allow_minting: bool,
    },
}

Structured representation of a call into a known Move script.

impl ScriptCall {
    pub fn encode(self) -> Script { .. }
    pub fn decode(&Script) -> Option<ScriptCall> { .. }
}

Variants

AddCurrencyToAccount

Add a Currency balance to account, which will enable account to send and receive Libra<Currency>. Aborts with NOT_A_CURRENCY if Currency is not an accepted currency type in the Libra system Aborts with LibraAccount::ADD_EXISTING_CURRENCY if the account already holds a balance in Currency.

Fields of AddCurrencyToAccount

currency: TypeTag

AddRecoveryRotationCapability

Add the KeyRotationCapability for to_recover_account to the RecoveryAddress resource under recovery_address.

Aborts

Aborts with LibraAccount::EKEY_ROTATION_CAPABILITY_ALREADY_EXTRACTED if account has already delegated its KeyRotationCapability.

Aborts with RecoveryAddress:ENOT_A_RECOVERY_ADDRESS if recovery_address does not have a RecoveryAddress resource.

Aborts with RecoveryAddress::EINVALID_KEY_ROTATION_DELEGATION if to_recover_account and recovery_address do not belong to the same VASP.

Fields of AddRecoveryRotationCapability

recovery_address: AccountAddress

AddToScriptAllowList

Append the hash to script hashes list allowed to be executed by the network.

Fields of AddToScriptAllowList

hash: Vec<u8>sliding_nonce: u64

AddValidatorAndReconfigure

Add new_validator to the validator set. Fails if the new_validator address is already in the validator set or does not have a ValidatorConfig resource stored at the address. Emits a NewEpochEvent.

Fields of AddValidatorAndReconfigure

sliding_nonce: u64validator_name: Vec<u8>validator_address: AccountAddress

Burn

Permanently destroy the Tokens stored in the oldest burn request under the Preburn resource. This will only succeed if account has a MintCapability<Token>, a Preburn<Token> resource exists under preburn_address, and there is a pending burn request. sliding_nonce is a unique nonce for operation, see sliding_nonce.move for details

Fields of Burn

token: TypeTagsliding_nonce: u64preburn_address: AccountAddress

BurnTxnFees

Burn transaction fees that have been collected in the given currency and relinquish to the association. The currency must be non-synthetic.

Fields of BurnTxnFees

coin_type: TypeTag

CancelBurn

Cancel the oldest burn request from preburn_address and return the funds. Fails if the sender does not have a published BurnCapability<Token>.

Fields of CancelBurn

token: TypeTagpreburn_address: AccountAddress

CreateChildVaspAccount

Create a ChildVASP account for sender parent_vasp at child_address with a balance of child_initial_balance in CoinType and an initial authentication_key auth_key_prefix | child_address. If add_all_currencies is true, the child address will have a zero balance in all available currencies in the system. This account will a child of the transaction sender, which must be a ParentVASP.

Aborts

The transaction will abort:

If parent_vasp is not a parent vasp with error: Roles::EINVALID_PARENT_ROLE

If child_address already exists with error: Roles::EROLE_ALREADY_ASSIGNED

If parent_vasp already has 256 child accounts with error: VASP::ETOO_MANY_CHILDREN

If CoinType is not a registered currency with error: LibraAccount::ENOT_A_CURRENCY

If parent_vasp's withdrawal capability has been extracted with error: LibraAccount::EWITHDRAWAL_CAPABILITY_ALREADY_EXTRACTED

If parent_vasp doesn't hold CoinType and child_initial_balance > 0 with error: LibraAccount::EPAYER_DOESNT_HOLD_CURRENCY

If parent_vasp doesn't at least child_initial_balance of CoinType in its account balance with error: LibraAccount::EINSUFFICIENT_BALANCE

Fields of CreateChildVaspAccount

coin_type: TypeTagchild_address: AccountAddressauth_key_prefix: Vec<u8>add_all_currencies: boolchild_initial_balance: u64

CreateDesignatedDealer

Create an account with the DesignatedDealer role at addr with authentication key auth_key_prefix | addr and a 0 balance of type Currency. If add_all_currencies is true, 0 balances for all available currencies in the system will also be added. This can only be invoked by an account with the TreasuryCompliance role.

Fields of CreateDesignatedDealer

currency: TypeTagsliding_nonce: u64addr: AccountAddressauth_key_prefix: Vec<u8>human_name: Vec<u8>add_all_currencies: bool

CreateParentVaspAccount

Create an account with the ParentVASP role at address with authentication key auth_key_prefix | new_account_address and a 0 balance of type currency. If add_all_currencies is true, 0 balances for all available currencies in the system will also be added. This can only be invoked by an Association account. sliding_nonce is a unique nonce for operation, see sliding_nonce.move for details.

Fields of CreateParentVaspAccount

coin_type: TypeTagsliding_nonce: u64new_account_address: AccountAddressauth_key_prefix: Vec<u8>human_name: Vec<u8>add_all_currencies: bool

CreateRecoveryAddress

Extract the KeyRotationCapability for recovery_account and publish it in a RecoveryAddress resource under account.

Aborts

Aborts with LibraAccount::EKEY_ROTATION_CAPABILITY_ALREADY_EXTRACTED if account has already delegated its KeyRotationCapability.

Aborts with RecoveryAddress::ENOT_A_VASP if account is not a ParentVASP or ChildVASP

Fields of CreateRecoveryAddress

CreateValidatorAccount

Create a validator account at new_validator_address with auth_key_prefixand human_name.

Fields of CreateValidatorAccount

sliding_nonce: u64new_account_address: AccountAddressauth_key_prefix: Vec<u8>human_name: Vec<u8>

CreateValidatorOperatorAccount

Create a validator operator account at new_validator_address with auth_key_prefixand human_name.

Fields of CreateValidatorOperatorAccount

sliding_nonce: u64new_account_address: AccountAddressauth_key_prefix: Vec<u8>human_name: Vec<u8>

FreezeAccount

Freeze account address. Initiator must be authorized. sliding_nonce is a unique nonce for operation, see sliding_nonce.move for details.

Fields of FreezeAccount

sliding_nonce: u64to_freeze_account: AccountAddress

MintLbr

Mint amount_lbr LBR from the sending account's constituent coins and deposits the resulting LBR into the sending account.

Fields of MintLbr

amount_lbr: u64

PeerToPeerWithMetadata

Transfer amount coins of type Currency from payer to payee with (optional) associated metadata and an (optional) metadata_signature on the message metadata | Signer::address_of(payer) | amount | DualAttestation::DOMAIN_SEPARATOR. The metadata and metadata_signature parameters are only required if amount >= DualAttestation::get_cur_microlibra_limit LBR and payer and payee are distinct VASPs. However, a transaction sender can opt in to dual attestation even when it is not required (e.g., a DesignatedDealer -> VASP payment) by providing a non-empty metadata_signature. Standardized metadata LCS format can be found in libra_types::transaction::metadata::Metadata.

Events

When this script executes without aborting, it emits two events: SentPaymentEvent { amount, currency_code = Currency, payee, metadata } on payer's LibraAccount::sent_events handle, and ReceivedPaymentEvent { amount, currency_code = Currency, payer, metadata } on payee's LibraAccount::received_events handle.

Common Aborts

These aborts can in occur in any payment.

Aborts with LibraAccount::EINSUFFICIENT_BALANCE if amount is greater than payer's balance in Currency.

Aborts with LibraAccount::ECOIN_DEPOSIT_IS_ZERO if amount is zero.

Aborts with LibraAccount::EPAYEE_DOES_NOT_EXIST if no account exists at the address payee.

Aborts with LibraAccount::EPAYEE_CANT_ACCEPT_CURRENCY_TYPE if an account exists at payee, but it does not accept payments in Currency.

Dual Attestation Aborts

These aborts can occur in any payment subject to dual attestation.

Aborts with DualAttestation::EMALFORMED_METADATA_SIGNATURE if metadata_signature's is not 64 bytes.

Aborts with DualAttestation:EINVALID_METADATA_SIGNATURE if metadata_signature does not verify on the message metadata | payer | value | DOMAIN_SEPARATOR using the compliance_public_key published in the payee's DualAttestation::Credential resource.

Other Aborts

These aborts should only happen when payer or payee have account limit restrictions or have been frozen by Libra administrators.

Aborts with LibraAccount::EWITHDRAWAL_EXCEEDS_LIMITS if payer has exceeded their daily withdrawal limits.

Aborts with LibraAccount::EDEPOSIT_EXCEEDS_LIMITS if payee has exceeded their daily deposit limits.

Aborts with LibraAccount::EACCOUNT_FROZEN if payer's account is frozen.

Fields of PeerToPeerWithMetadata

currency: TypeTagpayee: AccountAddressamount: u64metadata: Vec<u8>metadata_signature: Vec<u8>

Preburn

Preburn amount Tokens from account. This will only succeed if account already has a published Preburn<Token> resource.

Fields of Preburn

token: TypeTagamount: u64

PublishSharedEd25519PublicKey

(1) Rotate the authentication key of the sender to public_key (2) Publish a resource containing a 32-byte ed25519 public key and the rotation capability of the sender under the sender's address. Aborts if the sender already has a SharedEd25519PublicKey resource. Aborts if the length of new_public_key is not 32.

Fields of PublishSharedEd25519PublicKey

public_key: Vec<u8>

RegisterValidatorConfig

Set validator's config locally. Does not emit NewEpochEvent, the config is NOT changed in the validator set.

Fields of RegisterValidatorConfig

validator_account: AccountAddressconsensus_pubkey: Vec<u8>validator_network_identity_pubkey: Vec<u8>validator_network_address: Vec<u8>fullnodes_network_identity_pubkey: Vec<u8>fullnodes_network_address: Vec<u8>

RemoveValidatorAndReconfigure

Removes a validator from the validator set. Fails if the validator_address is not in the validator set. Emits a NewEpochEvent.

Fields of RemoveValidatorAndReconfigure

sliding_nonce: u64validator_name: Vec<u8>validator_address: AccountAddress

RotateAuthenticationKey

Rotate the sender's authentication key to new_key. new_key should be a 256 bit sha3 hash of an ed25519 public key.

Aborts with LibraAccount::EKEY_ROTATION_CAPABILITY_ALREADY_EXTRACTED if the KeyRotationCapability for account has already been extracted.

Aborts with 0 if the key rotation capability held by the account doesn't match the sender's address.

Aborts with LibraAccount::EMALFORMED_AUTHENTICATION_KEY if the length of new_key != 32.

Fields of RotateAuthenticationKey

new_key: Vec<u8>

RotateAuthenticationKeyWithNonce

Rotate account's authentication key to new_key. new_key should be a 256 bit sha3 hash of an ed25519 public key. This script also takes sliding_nonce, as a unique nonce for this operation. See sliding_nonce.move for details.

Fields of RotateAuthenticationKeyWithNonce

sliding_nonce: u64new_key: Vec<u8>

RotateAuthenticationKeyWithNonceAdmin

Rotate account's authentication key to new_key. new_key should be a 256 bit sha3 hash of an ed25519 public key. This script also takes sliding_nonce, as a unique nonce for this operation. See sliding_nonce.move for details.

Fields of RotateAuthenticationKeyWithNonceAdmin

sliding_nonce: u64new_key: Vec<u8>

RotateAuthenticationKeyWithRecoveryAddress

Rotate the authentication key of account to new_key using the KeyRotationCapability stored under recovery_address.

Aborts

Aborts with RecoveryAddress::ENOT_A_RECOVERY_ADDRESS if recovery_address does not have a RecoveryAddress resource

Aborts with RecoveryAddress::ECANNOT_ROTATE_KEY if account is not recovery_address or to_recover.

Aborts with LibraAccount::EMALFORMED_AUTHENTICATION_KEY if new_key is not 32 bytes.

Aborts with RecoveryAddress::ECANNOT_ROTATE_KEY if account has not delegated its KeyRotationCapability to recovery_address.

Fields of RotateAuthenticationKeyWithRecoveryAddress

recovery_address: AccountAddressto_recover: AccountAddressnew_key: Vec<u8>

RotateDualAttestationInfo

Rotate account's base URL to new_url and its compliance public key to new_key. Aborts if account is not a ParentVASP or DesignatedDealer Aborts if new_key is not a well-formed public key

Fields of RotateDualAttestationInfo

new_url: Vec<u8>new_key: Vec<u8>

RotateSharedEd25519PublicKey

(1) Rotate the public key stored in account's SharedEd25519PublicKey resource to new_public_key (2) Rotate the authentication key using the capability stored in account's SharedEd25519PublicKey to a new value derived from new_public_key Aborts if account does not have a SharedEd25519PublicKey resource. Aborts if the length of new_public_key is not 32.

Fields of RotateSharedEd25519PublicKey

public_key: Vec<u8>

SetValidatorConfigAndReconfigure

Set validator's config and updates the config in the validator set. NewEpochEvent is emitted.

Fields of SetValidatorConfigAndReconfigure

validator_account: AccountAddressconsensus_pubkey: Vec<u8>validator_network_identity_pubkey: Vec<u8>validator_network_address: Vec<u8>fullnodes_network_identity_pubkey: Vec<u8>fullnodes_network_address: Vec<u8>

SetValidatorOperator

Set validator's operator

Fields of SetValidatorOperator

operator_name: Vec<u8>operator_account: AccountAddress

SetValidatorOperatorWithNonceAdmin

Set validator operator as 'operator_account' of validator owner 'account' (via Admin Script). operator_name should match expected from operator account. This script also takes sliding_nonce, as a unique nonce for this operation. See Sliding_nonce.move for details.

Fields of SetValidatorOperatorWithNonceAdmin

sliding_nonce: u64operator_name: Vec<u8>operator_account: AccountAddress

TieredMint

Mint 'mint_amount' to 'designated_dealer_address' for 'tier_index' tier. Max valid tier index is 3 since there are max 4 tiers per DD. Sender should be treasury compliance account and receiver authorized DD. sliding_nonce is a unique nonce for operation, see sliding_nonce.move for details.

Fields of TieredMint

coin_type: TypeTagsliding_nonce: u64designated_dealer_address: AccountAddressmint_amount: u64tier_index: u64

UnfreezeAccount

Unfreeze account address. Initiator must be authorized. sliding_nonce is a unique nonce for operation, see sliding_nonce.move for details.

Fields of UnfreezeAccount

sliding_nonce: u64to_unfreeze_account: AccountAddress

UnmintLbr

Unmints amount_lbr LBR from the sending account into the constituent coins and deposits the resulting coins into the sending account.

Fields of UnmintLbr

amount_lbr: u64

UpdateDualAttestationLimit

Update the dual attesation limit to new_micro_lbr_limit.

Fields of UpdateDualAttestationLimit

sliding_nonce: u64new_micro_lbr_limit: u64

UpdateExchangeRate

Update the on-chain exchange rate to LBR for the given currency to be given by new_exchange_rate_numerator/new_exchange_rate_denominator.

Fields of UpdateExchangeRate

currency: TypeTagsliding_nonce: u64new_exchange_rate_numerator: u64new_exchange_rate_denominator: u64

UpdateLibraVersion

Update Libra version. sliding_nonce is a unique nonce for operation, see sliding_nonce.move for details.

Fields of UpdateLibraVersion

sliding_nonce: u64major: u64

UpdateMintingAbility

Allows--true--or disallows--false--minting of currency based upon allow_minting.

Fields of UpdateMintingAbility

currency: TypeTagallow_minting: bool

Implementations

impl ScriptCall

pub fn encode(self) -> Script

Build a Libra Script from a structured object ScriptCall.

pub fn decode(script: &Script) -> Option<ScriptCall>

Try to recognize a Libra Script and convert it into a structured object ScriptCall.