Function transaction_builder_generated::stdlib::encode_update_account_limit_definition_script

pub fn encode_update_account_limit_definition_script(
    coin_type: TypeTag, 
    limit_address: AccountAddress, 
    sliding_nonce: u64, 
    new_max_inflow: u64, 
    new_max_outflow: u64, 
    new_max_holding_balance: u64, 
    new_time_period: u64
) -> Script

Optionally update thresholds of max balance, inflow, outflow for any limits-bound accounts with their limits defined at limit_address. Limits are defined in terms of base (on-chain) currency units for CoinType. If a new threshold is 0, that particular config does not get updated. sliding_nonce is a unique nonce for operation, see SlidingNonce.move for details.